requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
httpx==0.28.1
python-dotenv==1.1.0
diskcache==5.6.3
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.1.1
slowapi==0.1.9